pyproject.toml
[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[project]
name = "smartwiki2"
version = "0.1.0"
description = "Control program for a Karpathy-style external wiki vault."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "openai>=1.40.0",
  "pydantic-settings>=2.5.0",
  "python-frontmatter>=1.1.0",
  "PyYAML>=6.0.1",
  "watchdog>=4.0.0",
  "httpx>=0.27.0",
  "pypdf",
  "python-docx",
  "python-pptx",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
